refactor(ema): Move study task models to folder
