+++ test/src/ema_db/datasources/firebase_datasource_test.dart
@@ -3,8 +3,8 @@ import 'package:fake_cloud_firestore/fake_cloud_firestore.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mdigit_span_tasks_ema/src/ema_db/datasources/firebase_datasource.dart';
 import 'package:mdigit_span_tasks_ema/src/ema_db/models/ema_model.dart';
-import 'package:mdigit_span_tasks_ema/src/ema_db/study_task/survey/survey_item.dart';
-import 'package:mdigit_span_tasks_ema/src/ema_db/study_task/task_item.dart';
+import 'package:mdigit_span_tasks_ema/src/ema_db/study_task/models/survey/survey_item.dart';
+import 'package:mdigit_span_tasks_ema/src/ema_db/study_task/models/task_item.dart';
 
 void main() {
   late FirebaseDataSource db;
